% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_walking.R
\docType{data}
\name{example_walking}
\alias{example_walking}
\title{Example Walking Data from IU (taken from \code{adeptdata})}
\format{
\subsection{\code{example_walking}}{

A \code{tibble} with 20653 rows and 2 columns, from left wrist, for ID
\code{id86237981}:
\describe{
\item{time}{Time from start, in seconds, a numeric}
\item{vm}{The vector magnitude signal}
}
}
}
\usage{
example_walking
}
\description{
Example Walking Data from IU (taken from \code{adeptdata})
}
\keyword{datasets}
